backend/requirements.txt
django
djangorestframework

psycopg2-binary

pyjwt >=1.7.1,<1.7.2

gunicorn==19.9.0